Added 4 Routes
Added routes to the Mechanism, Reaction, species, and Tag Mechanism Controller.

For Mechanism it is to get all mechanisms for a family uuid.

For Tag Mechanism its to get all tag mechanisms for a mechanism uuid.

For reactions and species its to get all the ones associated with a tag mechanism ID.